requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

pyoqs==0.10.0

cryptography==42.0.5
pydantic==2.6.4
